[#151]refactor: flowControl 패키지 생성 후 pass 이동
